Bug 1269740: Convert http to https URLs
This hacks the local copy of docson, as well as the bits of code parsing
data-doc-ref and data-doc-schema, to always load from https URLs, even
if the given URL is http.  This assumes that
https://schemas.taskcluster.net and https://references.taskcluster.net
are the same as their non-https cousins, and that all ref/schema URLs
are from one of those two sites.