Update a few inconsistent links in MD files.
There are 3 cases of links in the markdown links containing traling slashes. In
2 cases this results in broken links:
- tutorial/authenticate.md has a broken link to APIs
- manual/execution/workers/index.md has a broken link to worker interaction.

In another case this did not result in a broken link:
- manual/apis/roles.md links to namespaces.

It appears pages inside devel/ will redirect if issued at <base>/devel/page.
Which baffles me. However I figured this link should also be made consistent
with the absolute addressing used in the docs, so I've updated it also.