Remove things from best-practices document
As discussed in our 2018 Berlin work week, this commit removes sections
of this document which have disagreement and weren't added to the
document through a full consensus.  This is intended to be a lightweight
removal.  If these sections are desired to be in the document, we should
open a new PR to build full consensus around these issues.